worker: refactor start-attempt -> channels/attempt
