--- packages/ws-worker/src/channels/attempts
@@ -1,7 +1,0 @@
-// @ts-ignore
-
-// all the attempt stuff goes here
-// join the attempt channel
-// do we pull the attempt data here? I don't think so really tbh
-// it shoud prob be in execute, which is kinda annoying ebcause executeis designed to take a plan
-// i guess it just takes a cahannel and it pulls
